TestSuite Improvements. All Models in test classes  moved to Models folder.
