+++ MongoFramework.Tests/Mapping/Processors/NestedProperty/NestedPropertyProcessorTests.cs
@@ -2,20 +2,10 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using MongoDB.Bson.Serialization;
 using MongoFramework.Infrastructure.Mapping.Processors;
+using MongoFramework.Tests.Models;
 
 namespace MongoFramework.Tests.Mapping.Processors.NestedProperty
 {
-	public class ParentModel
-	{
-		public string Id { get; set; }
-		public NestedModel Model { get; set; }
-	}
-
-	public class NestedModel
-	{
-		public string HelloWorld { get; set; }
-	}
-
 	[TestClass]
 	public class NestedPropertyProcessorTests
 	{
