+++ MongoFramework.Tests/Models/ChangeTrackerModel.cs
@@ -1,0 +1,8 @@
+﻿namespace MongoFramework.Tests.Models
+{
+	public class ChangeTrackerModel
+	{
+		public string Id { get; set; }
+		public string Title { get; set; }
+	}
+}
