+++ MongoFramework.Tests/Models/ChildTestModel.cs
@@ -1,0 +1,9 @@
+﻿using MongoFramework.Tests.Mapping.Processors.Hierarchy;
+
+namespace MongoFramework.Tests.Models
+{
+	public class ChildTestModel : ParentTestModel
+	{
+		public string DeclaredProperty { get; set; }
+	}
+}
