+++ MongoFramework.Tests/Models/ColumnAttributePropertyModel.cs
@@ -1,0 +1,11 @@
+﻿using System.ComponentModel.DataAnnotations.Schema;
+
+namespace MongoFramework.Tests.Models
+{
+	public class ColumnAttributePropertyModel
+	{
+		public string Id { get; set; }
+		[Column("CustomPropertyName")]
+		public string MyProperty { get; set; }
+	}
+}
