+++ MongoFramework.Tests/Models/CustomCollectionModel.cs
@@ -1,0 +1,9 @@
+﻿using System.ComponentModel.DataAnnotations.Schema;
+
+namespace MongoFramework.Tests.Models
+{
+	[Table("CustomCollection")]
+	public class CustomCollectionModel
+	{
+	}
+}
