+++ MongoFramework.Tests/Models/DefaultCollectionNameModel.cs
@@ -1,0 +1,6 @@
+﻿namespace MongoFramework.Tests.Models
+{
+	public class DefaultCollectionNameModel
+	{
+	}
+}
