+++ MongoFramework.Tests/Models/EntityIndexWriterTests.NoIndexModel.cs
@@ -1,0 +1,10 @@
+﻿namespace MongoFramework.Tests.Models
+{
+	public partial class EntityIndexWriterTests
+	{
+		public class NoIndexModel
+		{
+			public string Id { get; set; }
+		}
+	}
+}
