+++ MongoFramework.Tests/Models/EntityWriterModel.cs
@@ -1,0 +1,8 @@
+﻿namespace MongoFramework.Tests.Models
+{
+	public class EntityWriterModel
+	{
+		public string Id { get; set; }
+		public string Title { get; set; }
+	}
+}
