+++ MongoFramework.Tests/Models/GuidIdGeneratorTestModel.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.ComponentModel.DataAnnotations;
+
+namespace MongoFramework.Tests.Models
+{
+	public class GuidIdGeneratorTestModel
+	{
+		[Key]
+		public Guid MyCustomId { get; set; }
+	}
+}
