+++ MongoFramework.Tests/Models/IdByAttributeTestModel.cs
@@ -1,0 +1,10 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace MongoFramework.Tests.Models
+{
+	public class IdByAttributeTestModel
+	{
+		[Key]
+		public string MyCustomId { get; set; }
+	}
+}
