+++ MongoFramework.Tests/Models/IgnoreExtraElementsModel.cs
@@ -1,0 +1,10 @@
+﻿using MongoFramework.Attributes;
+
+namespace MongoFramework.Tests.Models
+{
+	[IgnoreExtraElements]
+	public class IgnoreExtraElementsModel
+	{
+		public string Id { get; set; }
+	}
+}
