+++ MongoFramework.Tests/Models/InnerNestedTraverseMappingModel.cs
@@ -1,0 +1,8 @@
+﻿namespace MongoFramework.Tests.Models
+{
+	public class InnerNestedTraverseMappingModel
+	{
+		public string InnerMostProperty { get; set; }
+		public TraverseMappingModel NestedRecursionType { get; set; }
+	}
+}
