+++ MongoFramework.Tests/Models/MappingLockModel.cs
@@ -1,0 +1,7 @@
+﻿namespace MongoFramework.Tests.Models
+{
+	public class MappingLockModel
+	{
+		public string Id { get; set; }
+	}
+}
