+++ MongoFramework.Tests/Models/NestedModel.cs
@@ -1,0 +1,7 @@
+﻿namespace MongoFramework.Tests.Models
+{
+	public class NestedModel
+	{
+		public string HelloWorld { get; set; }
+	}
+}
