+++ MongoFramework.Tests/Models/NestedTraverseMappingModel.cs
@@ -1,0 +1,10 @@
+﻿
+namespace MongoFramework.Tests.Models
+{
+	public class NestedTraverseMappingModel
+	{
+		public string PropertyOne { get; set; }
+		public int PropertyTwo { get; set; }
+		public InnerNestedTraverseMappingModel InnerModel { get; set; }
+	}
+}
