+++ MongoFramework.Tests/Models/NotMappedPropertiesModel.cs
@@ -1,0 +1,11 @@
+﻿using System.ComponentModel.DataAnnotations.Schema;
+
+namespace MongoFramework.Tests.Models
+{
+	public class NotMappedPropertiesModel
+	{
+		public string Id { get; set; }
+		[NotMapped]
+		public string NotMapped { get; set; }
+	}
+}
