+++ MongoFramework.Tests/Models/ObjectIdGeneratorTestModel.cs
@@ -1,0 +1,11 @@
+﻿using System.ComponentModel.DataAnnotations;
+using MongoDB.Bson;
+
+namespace MongoFramework.Tests.Models
+{
+	public class ObjectIdGeneratorTestModel
+	{
+		[Key]
+		public ObjectId MyCustomId { get; set; }
+	}
+}
