+++ MongoFramework.Tests/Models/ParentModel.cs
@@ -1,0 +1,8 @@
+﻿namespace MongoFramework.Tests.Models
+{
+	public class ParentModel
+	{
+		public string Id { get; set; }
+		public NestedModel Model { get; set; }
+	}
+}
