+++ MongoFramework.Tests/Models/ParentTestModel.cs
@@ -1,0 +1,7 @@
+﻿namespace MongoFramework.Tests.Models
+{
+	public class ParentTestModel
+	{
+		public string Id { get; set; }
+	}
+}
