+++ MongoFramework.Tests/Models/StringIdGeneratorTestModel.cs
@@ -1,0 +1,10 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace MongoFramework.Tests.Models
+{
+	public class StringIdGeneratorTestModel
+	{
+		[Key]
+		public string MyCustomId { get; set; }
+	}
+}
