+++ MongoFramework.Tests/MongoFramework.Tests.csproj
@@ -73,6 +73,36 @@
     <Compile Include="AsyncDbEntityWriterTests.cs" />
     <Compile Include="Bson\GetDifferencesTests.cs" />
     <Compile Include="Bson\HasDifferencesTests.cs" />
+    <Compile Include="Models\GuidIdGeneratorTestModel.cs" />
+    <Compile Include="Models\IdByAttributeTestModel.cs" />
+    <Compile Include="Models\ObjectIdGeneratorTestModel.cs" />
+    <Compile Include="Models\StringIdGeneratorTestModel.cs" />
+    <Compile Include="Models\ExtraElementsModel.cs" />
+    <Compile Include="Models\IgnoreExtraElementsModel.cs" />
+    <Compile Include="Models\ChildTestModel.cs" />
+    <Compile Include="Models\ParentTestModel.cs" />
+    <Compile Include="Models\ColumnAttributePropertyModel.cs" />
+    <Compile Include="Models\NotMappedPropertiesModel.cs" />
+    <Compile Include="Models\NestedModel.cs" />
+    <Compile Include="Models\ParentModel.cs" />
+    <Compile Include="Models\CustomCollectionAndSchemaModel.cs" />
+    <Compile Include="Models\CustomCollectionModel.cs" />
+    <Compile Include="Models\DefaultCollectionNameModel.cs" />
+    <Compile Include="Models\InnerNestedTraverseMappingModel.cs" />
+    <Compile Include="Models\MappingLockModel.cs" />
+    <Compile Include="Models\NestedTraverseMappingModel.cs" />
+    <Compile Include="Models\TraverseMappingModel.cs" />
+    <Compile Include="Models\EntityIndexWriterTests.IndexModel.cs" />
+    <Compile Include="Models\EntityIndexWriterTests.NoIndexModel.cs" />
+    <Compile Include="Models\CompoundIndexModel.cs" />
+    <Compile Include="Models\IndexNamingModel.cs" />
+    <Compile Include="Models\IndexSortOrderModel.cs" />
+    <Compile Include="Models\NestedIndexBaseModel.cs" />
+    <Compile Include="Models\NestedIndexChildModel.cs" />
+    <Compile Include="Models\UniqueConstraintModel.cs" />
+    <Compile Include="Models\EntityWriterModel.cs" />
+    <Compile Include="Models\ExtendedEntityWriterModel.cs" />
+    <Compile Include="Models\ChangeTrackerModel.cs" />
     <Compile Include="DefinitionHelpers\UpdateDefinition\UpdateDefinitionHelperTests.cs" />
     <Compile Include="Indexing\EntityIndexMapperTests.cs" />
     <Compile Include="Indexing\EntityIndexWriterTests.cs" />
@@ -82,8 +112,14 @@
     <Compile Include="Mapping\Processors\Hierarchy\HierarchyProcessorTests.cs" />
     <Compile Include="Mapping\Processors\MappedProperties\MappedPropertiesProcessorTests.cs" />
     <Compile Include="Mapping\Processors\NestedProperty\NestedPropertyProcessorTests.cs" />
+    <Compile Include="Models\MongoDbContextModel.cs" />
+    <Compile Include="MongoDbContextTestContext.cs" />
     <Compile Include="MongoDbContextTests.cs" />
     <Compile Include="MongoDbSetTests.cs" />
+    <Compile Include="Models\MongoDbSetValidationModel.cs" />
+    <Compile Include="Mutation\Mutators\CreatedDate\InvalidAttributeUseModel.cs" />
+    <Compile Include="Mutation\Mutators\CreatedDate\ValidAttributeUseModel.cs" />
+    <Compile Include="Mutation\Mutators\IncrementNumber\IncrementalEntity.cs" />
     <Compile Include="Mutation\Mutators\IncrementNumber\IncrementNumberMutatorTests.cs" />
     <Compile Include="Mutation\Mutators\UpdatedDate\UpdatedDateMutatorTests.cs" />
     <Compile Include="Models\ExtendedEntity.cs" />
@@ -92,6 +128,7 @@
     <Compile Include="Models\ListRelationItem.cs" />
     <Compile Include="Models\SingleRelation.cs" />
     <Compile Include="Mutation\Mutators\CreatedDate\CreatedDateMutatorTests.cs" />
+    <Compile Include="Mutation\Mutators\UpdatedDate\ValidAttributeUseModel.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TestConfiguration.cs" />
     <Compile Include="TestSetup.cs" />
