+++ MongoFramework.Tests/Mutation/Mutators/CreatedDate/ValidAttributeUseModel.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using MongoFramework.Attributes;
+
+namespace MongoFramework.Tests.Mutation.Mutators.CreatedDate
+{
+	public class ValidAttributeUseModel
+	{
+		public string Id { get; set; }
+
+		[CreatedDate]
+		public DateTime CreatedDate { get; set; }
+	}
+}
