+++ MongoFramework.Tests/Mutation/Mutators/IncrementNumber/IncrementNumberMutatorTests.cs
@@ -1,21 +1,10 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using MongoFramework.Attributes;
 using MongoFramework.Infrastructure.Mapping;
 using MongoFramework.Infrastructure.Mutation;
 using MongoFramework.Infrastructure.Mutation.Mutators;
 
 namespace MongoFramework.Tests.Mutation.Mutators.IncrementNumber
 {
-	public class IncrementalEntity
-	{
-		[IncrementNumber]
-		public int ByDefault { get; set; }
-		[IncrementNumber(true)]
-		public int ByUpdateOnly { get; set; }
-		[IncrementNumber(10)]
-		public int ByTen { get; set; }
-	}
-
 	[TestClass]
 	public class IncrementNumberMutatorTests
 	{
