+++ MongoFramework.Tests/Mutation/Mutators/UpdatedDate/UpdatedDateMutatorTests.cs
@@ -7,14 +7,6 @@ using MongoFramework.Infrastructure.Mutation.Mutators;
 
 namespace MongoFramework.Tests.Mutation.Mutators.UpdatedDate
 {
-	public class ValidAttributeUseModel
-	{
-		public string Id { get; set; }
-
-		[UpdatedDate]
-		public DateTime UpdatedDate { get; set; }
-	}
-
 	public class InvalidAttributeUseModel
 	{
 		public string Id { get; set; }
